[LCR-136] refactor: popupAccess에서 visitLog로 패키지명 변경
