--- src/main/java/com/lgcns/domain/popupAccess/domain/UserGender.java
@@ -1,14 +1,0 @@
-package com.lgcns.domain.popupAccess.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum UserGender {
-    MALE("남성"),
-    FEMALE("여성"),
-    ;
-
-    private final String gender;
-}
