--- src/main/java/com/lgcns/domain/popupAccess/repository/PopupEnterRepository.java
@@ -1,6 +1,0 @@
-package com.lgcns.domain.popupAccess.repository;
-
-import com.lgcns.domain.popupAccess.domain.PopupEnter;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PopupEnterRepository extends JpaRepository<PopupEnter, Long> {}
