--- src/main/java/com/lgcns/domain/popupAccess/service/PopupAccessService.java
@@ -1,7 +1,0 @@
-package com.lgcns.domain.popupAccess.service;
-
-import com.lgcns.domain.popupAccess.dto.request.PopupEnterCreateRequest;
-
-public interface PopupAccessService {
-    void createPopupEnter(Long popupId, PopupEnterCreateRequest request);
-}
