+++ src/main/resources/db/migration/V8__create_entrance_table.sql
@@ -1,5 +1,5 @@
-CREATE TABLE popup_enter (
-                             popup_enter_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+CREATE TABLE entrance (
+                             entrance_id BIGINT AUTO_INCREMENT PRIMARY KEY,
                              popup_id BIGINT NOT NULL,
                              gender VARCHAR(10) NOT NULL,
                              age_group INT NOT NULL,
